@chrisvest 2021-11-24
An alternative take is #8052. Not sure which one is best, but I'll point out that this version has fewer changes to the tests, and does away with weak references entirely.
I just noticed the benchmark, which would be good to copy over.