@franz1981 2021-11-24
Very happy that we're getting rid of weak ref really: anything that's not using it is a +100 from me :P