@chrisvest 2021-11-25
Got benchmark numbers.

Old version:

```
Benchmark                                                                Mode  Cnt     Score     Error   Units
RecyclerBenchmark.plainNew                                               avgt   20     7.048 ±   0.087   ns/op
RecyclerBenchmark.plainNew:·gc.alloc.rate                                avgt   20  7210.104 ±  88.082  MB/sec
RecyclerBenchmark.plainNew:·gc.alloc.rate.norm                           avgt   20    80.000 ±   0.001    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space                       avgt   20  7211.559 ± 118.469  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space.norm                  avgt   20    80.015 ±   0.764    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space                   avgt   20     0.059 ±   0.019  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space.norm              avgt   20     0.001 ±   0.001    B/op
RecyclerBenchmark.plainNew:·gc.count                                     avgt   20   851.000            counts
RecyclerBenchmark.plainNew:·gc.time                                      avgt   20   640.000                ms
RecyclerBenchmark.producerConsumer                                       avgt   20   350.694 ±  21.272   ns/op
RecyclerBenchmark.producerConsumer:consumer                              avgt   20   350.693 ±  21.272   ns/op
RecyclerBenchmark.producerConsumer:producer                              avgt   20   350.694 ±  21.272   ns/op
RecyclerBenchmark.producerConsumer:·gc.alloc.rate                        avgt   20    50.777 ±   2.966  MB/sec
RecyclerBenchmark.producerConsumer:·gc.alloc.rate.norm                   avgt   20    13.990 ±   0.807    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space               avgt   20    42.011 ±  57.813  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space.norm          avgt   20    11.477 ±  15.721    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space           avgt   20     0.151 ±   0.393  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space.norm      avgt   20     0.039 ±   0.102    B/op
RecyclerBenchmark.producerConsumer:·gc.count                             avgt   20     6.000            counts
RecyclerBenchmark.producerConsumer:·gc.time                              avgt   20    11.000                ms
RecyclerBenchmark.recyclerGetAndOrphan                                   avgt   20    11.686 ±   0.087   ns/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate                    avgt   20  6088.941 ±  44.558  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate.norm               avgt   20   112.000 ±   0.001    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space           avgt   20  6092.570 ±  90.371  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space.norm      avgt   20   112.064 ±   1.242    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space       avgt   20     0.153 ±   0.066  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space.norm  avgt   20     0.003 ±   0.001    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.count                         avgt   20   719.000            counts
RecyclerBenchmark.recyclerGetAndOrphan:·gc.time                          avgt   20   573.000                ms
RecyclerBenchmark.recyclerGetAndRecycle                                  avgt   20    16.333 ±   0.732   ns/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate                   avgt   20     0.001 ±   0.002  MB/sec
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate.norm              avgt   20    ≈ 10⁻⁵              B/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.count                        avgt   20       ≈ 0            counts
```

This version:

```
Benchmark                                                                Mode  Cnt     Score     Error   Units
RecyclerBenchmark.plainNew                                               avgt   20     6.878 ±   0.032   ns/op
RecyclerBenchmark.plainNew:·gc.alloc.rate                                avgt   20  7387.032 ±  34.000  MB/sec
RecyclerBenchmark.plainNew:·gc.alloc.rate.norm                           avgt   20    80.000 ±   0.001    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space                       avgt   20  7396.146 ±  69.736  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Eden_Space.norm                  avgt   20    80.098 ±   0.544    B/op
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space                   avgt   20     0.061 ±   0.023  MB/sec
RecyclerBenchmark.plainNew:·gc.churn.PS_Survivor_Space.norm              avgt   20     0.001 ±   0.001    B/op
RecyclerBenchmark.plainNew:·gc.count                                     avgt   20   873.000            counts
RecyclerBenchmark.plainNew:·gc.time                                      avgt   20   681.000                ms
RecyclerBenchmark.producerConsumer                                       avgt   20   354.463 ±  28.344   ns/op
RecyclerBenchmark.producerConsumer:consumer                              avgt   20   354.464 ±  28.346   ns/op
RecyclerBenchmark.producerConsumer:producer                              avgt   20   354.463 ±  28.342   ns/op
RecyclerBenchmark.producerConsumer:·gc.alloc.rate                        avgt   20    38.304 ±   2.483  MB/sec
RecyclerBenchmark.producerConsumer:·gc.alloc.rate.norm                   avgt   20    10.647 ±   0.650    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space               avgt   20    38.295 ±  52.115  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Eden_Space.norm          avgt   20    10.657 ±  14.570    B/op
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space           avgt   20     0.003 ±   0.008  MB/sec
RecyclerBenchmark.producerConsumer:·gc.churn.PS_Survivor_Space.norm      avgt   20     0.001 ±   0.002    B/op
RecyclerBenchmark.producerConsumer:·gc.count                             avgt   20     6.000            counts
RecyclerBenchmark.producerConsumer:·gc.time                              avgt   20    12.000                ms
RecyclerBenchmark.recyclerGetAndOrphan                                   avgt   20     9.557 ±   0.350   ns/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate                    avgt   20  5590.135 ± 201.291  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.alloc.rate.norm               avgt   20    84.000 ±   0.001    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space           avgt   20  5597.228 ± 219.528  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Eden_Space.norm      avgt   20    84.101 ±   0.973    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space       avgt   20     0.114 ±   0.057  MB/sec
RecyclerBenchmark.recyclerGetAndOrphan:·gc.churn.PS_Survivor_Space.norm  avgt   20     0.002 ±   0.001    B/op
RecyclerBenchmark.recyclerGetAndOrphan:·gc.count                         avgt   20   661.000            counts
RecyclerBenchmark.recyclerGetAndOrphan:·gc.time                          avgt   20   519.000                ms
RecyclerBenchmark.recyclerGetAndRecycle                                  avgt   20    20.349 ±   0.170   ns/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate                   avgt   20     0.001 ±   0.002  MB/sec
RecyclerBenchmark.recyclerGetAndRecycle:·gc.alloc.rate.norm              avgt   20    ≈ 10⁻⁴              B/op
RecyclerBenchmark.recyclerGetAndRecycle:·gc.count                        avgt   20       ≈ 0            counts
```

Summary:

* Very small performance improvement, except in `recyclerGetAndRecycle`, which regressed a bit.
* Allocation rate is generally down, despite more work done. I suspect much of this being because `DefaultHandle` objects are now smaller.